% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mdm.R
\name{annx}
\alias{annx}
\title{Expected value of annuity (matrix form)}
\usage{
annx(object, n = 10, x = 0, h = 1/12, delta = log(1.05), dis = 0,
  joint = FALSE, y = 0)
}
\arguments{
\item{object}{An object of class msm}

\item{n}{The number of years for the annuity}

\item{x}{The age of the individual}

\item{h}{The time step used in the approximation}

\item{delta}{The force of interest (continuously compounded interest rate)}

\item{dis}{Indicates that the annuity is discrete (with period h). By default, continuous annuities are calculated}

\item{joint}{Indicates whether the annuity depends on two lives}

\item{y}{The age of the second individual if the model is joint}
}
\value{
A matrix of expected values for annuities
}
\description{
Calculates the expected values of annuities whose values depend on transitioning from one state to another
}

