% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mdm.R
\name{Ax}
\alias{Ax}
\title{Expected value of insurance (matrix form)}
\usage{
Ax(object, n = 10, x = 0, h = 1/12, delta = log(1.05), joint = FALSE,
  y = 0)
}
\arguments{
\item{object}{An object of class msm}

\item{n}{The number of years for the insurance}

\item{x}{The age of the individual}

\item{h}{The time step used in the approximation}

\item{delta}{The force of interest (continuously compounded interest rate)}

\item{joint}{Indicates whether the insurance depends on two lives}

\item{y}{The age of the second individual if the model is joint}

\item{dis}{Indicates that the insurance is discrete (with period h). By default, continuous insurance are calculated}
}
\value{
A matrix of expected values for insurances
}
\description{
Calculates the expected values of annuities whose values depend on transitioning from one state to another
}

