% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDS.R
\name{BDS}
\alias{BDS}
\title{BDS}
\usage{
BDS(
  num_items = 18L,
  with_welcome = TRUE,
  with_training = TRUE,
  with_finish = TRUE,
  label = "BDS",
  feedback = BDS_feedback_with_score(),
  dict = BDS::BDS_dict
)
}
\arguments{
\item{num_items}{(Integer scalar) Number of items in the test.}

\item{with_welcome}{(Logical scalar) Whether to show a welcome page.}

\item{with_training}{(Logical scalar) Whether to include the training phase.}

\item{with_finish}{(Logical scalar) Whether to show a finished page.}

\item{label}{(Character scalar) Label to give the BDS results in the output file.}

\item{feedback}{(Function) Defines the feedback to give the participant
at the end of the test.}

\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
This function defines a BDS  module for incorporation into a
psychTestR timeline.
Use this function if you want to include the BDS in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For demoing the BDS, consider using \code{\link{BDS_demo}()}.
For a standalone implementation of the BDS,
consider using \code{\link{BDS_standalone}()}.
}
