% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{BDS_feedback_with_graph}
\alias{BDS_feedback_with_graph}
\title{BDS feedback (with graph)}
\usage{
BDS_feedback_with_graph(dict = BDS::BDS_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given textual and graphical feedback at the end of the test.
}
\examples{
\dontrun{
BDS_demo(feedback = BDS_feedback_with_score())}
}
