% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{BDS_standalone}
\alias{BDS_standalone}
\title{Standalone BDS}
\usage{
BDS_standalone(
  title = NULL,
  num_items = 18L,
  with_id = TRUE,
  feedback = BDS::BDS_feedback_with_graph(),
  with_welcome = TRUE,
  admin_password = "conifer",
  researcher_email = "longgold@gold.uc.ak",
  languages = c("EN", "DE"),
  dict = BDS::BDS_dict,
  validate_id = "auto",
  ...
)
}
\arguments{
\item{title}{(Scalar character) Title to display during testing.}

\item{num_items}{(Scalar integer) Number of items to be adminstered.}

\item{with_id}{(Logical scalar) Whether to show a ID page.}

\item{feedback}{(Function) Defines the feedback to give the participant
at the end of the test.}

\item{with_welcome}{(Logical scalar) Whether to show a welcome page.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}),
and German (\code{"DE"}).
The first language is selected by default}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{validate_id}{(Character scalar or closure) Function for validating IDs or string "auto" for default validation
which means ID should consist only of  alphanumeric characters.}

\item{...}{Further arguments to be passed to \code{\link{BDS}()}.}
}
\description{
This function launches a standalone testing session for the BDS
This can be used for data collection, either in the laboratory or online.
}
