% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bmsm.R
\name{Bmsm}
\alias{Bmsm}
\title{Bivariate Markov Switching Multifractal (BMSM) Model.}
\usage{
Bmsm(ret, kbar = 1, n = 252, para0 = NULL, s.err = TRUE)
}
\arguments{
\item{ret}{is a matrix of two returns.}

\item{kbar}{is the number of frequency components in the bivariate MSM model. Default is 1.}

\item{n}{is the number of trading days in a year, if data is on a daily frequency. Default is 252.
If data is monthly, then set n.vol to 12. If data is quarterly, then n.vol should be 4.}

\item{para0}{is the initial parameter values for optimization,
c(m01, m02, sigma1, sigma2, gamma-k, b, rho.e, labda, rho.m). Default is NULL.}

\item{s.err}{=T. If True, estimates standard error. Otherwise, no standard errors are estimated.
The computation of BMSM standard error is expensive. Hence, the option to skip it if not needed.}
}
\value{
a list consisting of:
\item{LL}{sum of loglikelihood values at optimum}
\item{LLs}{a vector of loglikelihood values at optimum}
\item{filtered}{a matrix of filtered probabilities}
\item{A}{the estimated transition matrix}
\item{g.m}{the estimated state values}
\item{optim.msg}{optimization message}
\item{optim.convergence}{optimization convergence indicator}
\item{optim.iter}{number of iterations}
\item{para]}{estimated parameter vector}
\item{se}{robust standard errors}
}
\description{
Estimates the parameters of a bivariate MSM model for a given pair of returns.
}
\note{
Use the \code{\link{summary.bmsmmodel}} function to print summary results of the BMSM model.
Use \code{\link{plot.bmsmmodel}} function to plot the fitted conditional volatilities, or plot(bmsmmodel) to plot the
 fitted  conditional variance. Use \code{\link{predict.bmsmmodel}} to get the predicted conditional
 variance, standard deviation, correlation and covariance.
}
\examples{
data("calvet2006returns")
ret <- as.matrix(calvet2006returns[,2:3])*100
fit <- Bmsm(ret)
fit <- Bmsm(ret, kbar=2, n=252, s.err=FALSE)
summary(fit)
plot(fit)

}
\author{
Waleem Alausa, \email{alausa.babs@gmail.com}
}
\references{
Calvet, L.E & Fisher, J.A (2006) Volatility comovement: a multifrequency approach
(\href{http://www.sciencedirect.com/science/article/pii/S0304-4076(05)00010-2}{PubMed})
}
\keyword{Markov}
\keyword{Multifractal,}
\keyword{Switching}
\keyword{clustering}
\keyword{volatility}
\keyword{volatility,}

