% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bmsm_check_lls2_para.R
\name{Bmsm_check_lls2_para}
\alias{Bmsm_check_lls2_para}
\title{Internal Routine for \code{\link{Bmsm}} Model.}
\usage{
Bmsm_check_lls2_para(x)
}
\arguments{
\item{x}{is a vector of \code{\link{Bmsm}} parameters.}
}
\value{
a 3 by 1 vector of \code{\link{Bmsm}} parameters.
}
\description{
Internal routine used by \code{\link{Bmsm_std_err}} model, to check that the
hessian and gradient calculations conform with the Bmsm parameter conditions.
}

