% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bmsm_clustermat.R
\name{Bmsm_clustermat}
\alias{Bmsm_clustermat}
\title{Volatility State Matrix for \code{\link{Bmsm}}(k) model.}
\usage{
Bmsm_clustermat(m01, m02, kbar)
}
\arguments{
\item{m01}{is the volatility state value for return series 1 in (1, 2].}

\item{m02}{is the volatility state value for return series 2 in (1, 2].}

\item{kbar}{is the number of frequency components in the \code{\link{Bmsm}}(k) model.}
}
\value{
a \eqn{4^k} by \eqn{2^k} grid of volatility components.
}
\description{
Calculates the matrix of volatility components for \code{\link{Bmsm}}(k) model.
}
\examples{
Bmsm_clustermat(1.5,1.7,2)

}

