% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bmsm_decompose.R
\name{Bmsm_decompose}
\alias{Bmsm_decompose}
\title{Internal Routine Used By \code{\link{Bmsm}} Model.}
\usage{
Bmsm_decompose(smoothed.p, m01, m02, kbar)
}
\arguments{
\item{smoothed.p}{is a matrix of smoothed probabilities from \code{\link{Bmsm}} Model.}

\item{m01}{is the volatility state value for return series 1 in (1, 2].}

\item{m02}{is the volatility state value for return series 2 in (1, 2].}

\item{kbar}{is the number of frequency components in the \code{\link{Bmsm}}(k) model.}
}
\value{
decomposed volatility components.
}
\description{
Internal Routine Used By \code{\link{Bmsm}} Model.
}

