% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bmsm_filtered2.R
\name{Bmsm_filtered2}
\alias{Bmsm_filtered2}
\title{Filtered Probability for \code{\link{Bmsm}}(k) model.}
\usage{
Bmsm_filtered2(para, kbar, dat, n.vol)
}
\arguments{
\item{para}{is a vector of parameters returned by \code{\link{Bmsm}}.}

\item{kbar}{is the number of frequency components in the \code{\link{Bmsm}}(k) model.}

\item{dat}{is a matrix of returns.}

\item{n.vol}{is the number of trading days in a year, if data is on a daily frequency. Default is 252.}
}
\value{
a list consisting of:
\item{LL}{sum of loglikelihood values at optimum}
\item{LLs}{a vector of loglikelihood values at optimum}
\item{filtered}{a matrix of filtered probabilities}
\item{A}{the estimated transition matrix}
\item{g.m}{the estimated state values}
}
\description{
Calculates filtered probability and log-likelihood values for \code{\link{Bmsm}}(k) model.
}
\examples{
data("calvet2006returns")
ret <- as.matrix(calvet2006returns[,2:3])*100
fit <- Bmsm(ret, kbar=2, n=252, s.err=FALSE)
x   <- Bmsm_filtered2(fit$para, 2, ret, 252)

}

