% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bmsm_grad.R
\name{Bmsm_grad}
\alias{Bmsm_grad}
\title{Numerical Derivative Function for \code{\link{Bmsm}}(k) model.}
\usage{
Bmsm_grad(func, arg.list)
}
\arguments{
\item{func}{is a function handle.}

\item{arg.list}{is a list of arguments for function "func".}
}
\value{
a  matrix of numerical derivatives.
}
\description{
Calculates the numerical derivative for \code{\link{Bmsm}}(k) model.
}

