% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bmsm_hessian_2_sided.R
\name{Bmsm_hessian_2_sided}
\alias{Bmsm_hessian_2_sided}
\title{Finite Difference Hessian for \code{\link{Bmsm}}(k) model.}
\usage{
Bmsm_hessian_2_sided(func, arg.list)
}
\arguments{
\item{func}{is a function handle.}

\item{arg.list}{is a list of arguments for function "func".}
}
\value{
a  matrix of numerical hessians.
}
\description{
Calculates 2-sided finite difference hessian for \code{\link{Bmsm}}(k) model.
}

