% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bmsm_parameter_check.R
\name{Bmsm_parameter_check}
\alias{Bmsm_parameter_check}
\title{Boundary Checks on \code{\link{Bmsm}} Model Parameters.}
\usage{
Bmsm_parameter_check(dat, kbar, x0, n)
}
\arguments{
\item{dat}{is a column matrix/ dataframe of returns.}

\item{kbar}{is the number of frequency components in the \code{\link{Msm}}(k) model.}

\item{x0}{is the initial parameter values passed to Msm.}

\item{n}{is the number of trading days in a year, if data is on a daily frequency. Default is 252.
If data is monthly, then set n.vol to 12. If data is quarterly, then n.vol should be 4.}
}
\value{
a list consisting of:
\item{dat}{is a column matrix of returns}
\item{kbar}{number of frequency components}
\item{start.value}{is the initial parameter values passed to Msm}
\item{lb1}{lower bound on stage 1 parameters}
\item{ub1}{upper bound on stage 1 parameters}
\item{lb2}{lower bound on stage 2 parameters}
\item{ub2}{upper bound on stage 2 parameters}
}
\description{
Performs boundary checks on \code{\link{Bmsm}} model parameter values.
}

