% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bmsm_predict.R
\name{Bmsm_predict}
\alias{Bmsm_predict}
\title{Prediction for \code{\link{Bmsm}} Model.}
\usage{
Bmsm_predict(smoothed.p, A, g.m, para, kbar, n.vol, h = NULL)
}
\arguments{
\item{smoothed.p}{is a matrix of smoothed/filtered state probabilities.}

\item{A}{is the transition matrix.}

\item{g.m}{is is a vector of possible \code{\link{Bmsm}} states.}

\item{para}{is a vector of \code{\link{Bmsm}} parameters.}

\item{kbar}{is the number of volatility components in \code{\link{Bmsm}}.}

\item{n.vol}{is the number of trading days in a year, if data is on a daily frequency. Default is 252.
If data is monthly, then set n.vol to 12. If data is quarterly, then n.vol should be 4.}

\item{h}{is an optional parameter. When h is null, the fitted volatility values are returned.
When h>0, then an h-step ahead volatility forecast is returned.}
}
\value{
a list consisting of:
\item{vol1}{fitted/forecasted volatility values for series 1}
\item{vol2}{fitted/forecasted volatility values for series 2}
\item{vol1.sq}{fitted/forecasted variance values for series 1}
\item{vol2.sq}{fitted/forecasted variance values for series 2}
\item{rho.t}{fitted/forecasted conditional correlation}
\item{covt}{fitted/forecasted conditional covariance}
}
\description{
Performs prediction after fitting a \code{\link{Bmsm}}(k) model.
}
\examples{
data("calvet2006returns")
ret <- as.matrix(calvet2006returns[,2:3])*100
fit <- Bmsm(ret, kbar=2, n=252, s.err=FALSE)
v=Bmsm_predict(fit$filtered.P, fit$A, fit$g.m, fit$para, fit$kbar, fit$n)

}

