% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bmsm_stage1_likelihood.R
\name{Bmsm_stage1_likelihood}
\alias{Bmsm_stage1_likelihood}
\title{Combined Univariate Log-Likelihood  For \code{\link{Bmsm}}(k) model.}
\usage{
Bmsm_stage1_likelihood(para, dat, kbar, n.vol)
}
\arguments{
\item{para}{is a 6 elements vector of parameters [m01, m02, sigma1, sigma2, gamma_k, b].}

\item{dat}{is a matrix of returns.}

\item{kbar}{is the number of frequency components in the \code{\link{Bmsm}}(k) model.}

\item{n.vol}{is the number of trading days in a year, if data is on a daily frequency. Default is 252.
If data is monthly, then set n.vol to 12. If data is quarterly, then n.vol should be 4.}
}
\value{
sum of the combined log-likelihood: L1 + L2
}
\description{
Calculates sum of the combined univariate log-likelihood values for \code{\link{Bmsm}}(k) model.
This is the first stage log-likelihood estimation for a \code{\link{Bmsm}} model.
This stage estimates the first six parameters [m01, m02, sigma1, sigma2, gamma_k, b].
}
\examples{
data("calvet2006returns")
ret <- as.matrix(calvet2006returns[,2:3])*100
L <- Bmsm_stage1_likelihood(c(1.4330, 1.5768, 5.0433, 7.6284, 0.7299, 8.2716), ret, 2, 252)

}

