% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bmsm_stage1hess_LL.R
\name{Bmsm_stage1hess_LL}
\alias{Bmsm_stage1hess_LL}
\title{Internal Routine Used By \code{\link{Bmsm}} Model.}
\usage{
Bmsm_stage1hess_LL(para, kbar, dat, n.vol)
}
\arguments{
\item{para}{is a vector of stage 1 \code{\link{Bmsm}} parameters.}

\item{kbar}{is the number of frequency components in the \code{\link{Bmsm}}(k) model.}

\item{dat}{is a matrix of returns.}

\item{n.vol}{is the number of trading days in a year, if data is on a daily frequency. Default is 252.
If data is monthly, then set n.vol to 12. If data is quarterly, then n.vol should be 4.}
}
\value{
Log-likelihood values.
}
\description{
Internal routine used by \code{\link{Bmsm}} model to calculate standard errors.
}

