% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bmsm_stage2_likelihood2.R
\name{Bmsm_stage2_likelihood2}
\alias{Bmsm_stage2_likelihood2}
\title{Log-Likelihood  For \code{\link{Bmsm}}(k) model.}
\usage{
Bmsm_stage2_likelihood2(para, kbar, dat, para1, n.vol)
}
\arguments{
\item{para}{is a 3 elements vector of parameters [rhoe, lamda, rho.m].}

\item{kbar}{is the number of frequency components in the \code{\link{Bmsm}}(k) model.}

\item{dat}{is a matrix of returns.}

\item{para1}{is a vector of first stage parameters [m01, m02, sigma1, sigma2, gamma_k, b].}

\item{n.vol}{is the number of trading days in a year, if data is on a daily frequency. Default is 252.
If data is monthly, then set n.vol to 12. If data is quarterly, then n.vol should be 4.}
}
\value{
sum of log-likelihood: L
}
\description{
Calculates sum of log-likelihood values for \code{\link{Bmsm}}(k) model.
This is the second stage log-likelihood estimation for a \code{\link{Bmsm}} model.
This stage estimates the last three parameters [rhoe, lamda, rho.m].
}
\examples{
data("calvet2006returns")
ret <- as.matrix(calvet2006returns[,2:3])*100
para1 <- c(1.4330, 1.5768, 5.0433, 7.6284, 0.7299, 8.2716)
para <- c(0.1090, 0.4421, 0.8823)
L <- Bmsm_stage2_likelihood2(para, 2, ret, para1, 252)

}

