% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bmsm_stage2_LLs.R
\name{Bmsm_stage2_LLs}
\alias{Bmsm_stage2_LLs}
\title{Internal Routine Used By \code{\link{Bmsm}} Model.}
\usage{
Bmsm_stage2_LLs(para, kbar, dat, n.vol, para1)
}
\arguments{
\item{para}{is a vector of stage 1 \code{\link{Bmsm}} parameters.}

\item{kbar}{is the number of frequency components in the \code{\link{Bmsm}}(k) model.}

\item{dat}{is a matrix of returns.}

\item{n.vol}{is the number of trading days in a year, if data is on a daily frequency. Default is 252.
If data is monthly, then set n.vol to 12. If data is quarterly, then n.vol should be 4.}

\item{para1}{is a vector of first stage parameters [m01, m02, sigma1, sigma2, gamma_k, b].}
}
\value{
Log-likelihood values.
}
\description{
Internal routine used by \code{\link{Bmsm}} model to calculate standard errors.
}

