% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bmsm_states.R
\name{Bmsm_states}
\alias{Bmsm_states}
\title{Volatility State Vector For \code{\link{Bmsm}} model.}
\usage{
Bmsm_states(m01, m02, kbar)
}
\arguments{
\item{m01}{is the state variable value for series 1.}

\item{m02}{is the state variable value for series 2.}

\item{kbar}{is the number of frequency components in the \code{\link{Bmsm}}(k) model.}
}
\value{
a  2-by-\eqn{4^k} state matrix.
}
\description{
Calculates all possible \eqn{4^k} state values for an  \code{\link{Bmsm}}(k) model.
}
\examples{
s <- Bmsm_states(1.5, 1.7, 2)

}

