% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bmsm_std_err.R
\name{Bmsm_std_err}
\alias{Bmsm_std_err}
\title{Standard Error for \code{\link{Bmsm}} Model.}
\usage{
Bmsm_std_err(para, kbar, ret, n.vol)
}
\arguments{
\item{para}{is a vector of parameters returned by \code{\link{Bmsm}}.}

\item{kbar}{is the number of frequency components in the \code{\link{Bmsm}}(k) model.}

\item{ret}{is a column matrix of returns.}

\item{n.vol}{is the number of trading days in a year, if data is on a daily frequency. Default is 252.}
}
\value{
a 9 by 1 vector of standard error values.
}
\description{
Calculates standard error for \code{\link{Bmsm}} model.
}

