% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bmsm_Transition.R
\name{Bmsm_Transition}
\alias{Bmsm_Transition}
\title{Internal Routine Used By \code{\link{Bmsm}} Model.}
\usage{
Bmsm_Transition(gama, lamda, rho.m)
}
\arguments{
\item{gama}{is the transition probability of the kth frequency component and must be in the range (0,1).}

\item{lamda}{in [0, 1] is the unconditional correlation between volatility arrivals.}

\item{rho.m}{in [-1, 1] is the correlation between volatility components \eqn{M_a} and \eqn{M_b}}
}
\value{
4-by-4 matrix.
}
\description{
Internal routine used by \code{\link{Bmsm}} model to calculate transition matrix.
}

