% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calvet2004data}
\alias{calvet2004data}
\title{Data from "How to Forecast Long-Run Volatility Paper"}
\format{An object of class \code{"data.frame"}.}
\source{
\href{https://fred.stlouisfed.org}{Federal Reserve Economic Data}
}
\usage{
data(calvet2004data)
}
\description{
Data from "How to Forecast Long-Run Volatility:Regime Switching and the
Estimation of Multifractal Processes".
Daily exchange rate data for the deutsche mark (dm), Japanese yen (ja),
British pound (uk), and Canadian dollar (ca), all against the U.S. dollar.
The data consist of daily prices (buying rates for wire transfers at 12:00 P.M. Eastern time)
 reported at noon by the Federal Reserve Bank of New York .
}
\examples{
data(calvet2004data)
caret <- as.matrix(na.omit(calvet2004data$caret))
\donttest{summary(caret)}
}
\references{
Calvet, L.E & Fisher, J.A (2004) How to Forecast Long-Run Volatility:Regime Switching and the
Estimation of Multifractal Processes
(\href{https://academic.oup.com/jfec/article-abstract/2/1/49/960710/How-to-Forecast-Long-Run-Volatility-Regime}{PubMed})
}
\keyword{datasets}

