% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calvet2006returns}
\alias{calvet2006returns}
\title{Data from "Volatility comovement: a multifrequency approach" Paper.}
\format{A data frame with 7635 rows and 4 variables:
\describe{
  \item{Date}{Daily date}
  \item{CanRet}{Log returns of Canada / U.S. Foreign Exchange Rate, Canadian Dollars to One U.S. Dollar, Daily}
  \item{JARet}{Log returns of Japan / U.S. Foreign Exchange Rate, Japanese Yen to One U.S. Dollar, Daily}
  \item{UKRet}{Log returns of U.K. / U.S. Foreign Exchange Rate, U.S. Dollars to One British Pound, Daily}
  }}
\source{
\href{https://fred.stlouisfed.org}{Federal Reserve Economic Data}
}
\usage{
data(calvet2006returns)
}
\description{
Data from "Volatility comovement: a multifrequency approach".
Daily exchange rate data for the Japanese yen (ja),
British pound (uk), and Canadian dollar (ca), all against the U.S. dollar.
The data consist of daily prices (buying rates for wire transfers at 12:00 P.M. Eastern time)
 reported at noon by the Federal Reserve Bank of New York .
}
\examples{
data(calvet2006returns)
caret <- as.matrix(calvet2006returns$CanRet)
\donttest{summary(caret)}
}
\references{
Calvet, L.E & Fisher, J.A (2006) Volatility comovement: a multifrequency approach
(\href{http://www.sciencedirect.com/science/article/pii/S0304-4076(05)00010-2}{PubMed})
}
\keyword{datasets}

