% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{cfm}
\alias{cfm}
\title{Constant Force of Mortality [msm]}
\format{\preformatted{Formal class 'msm' [package "MSM"] with 3 slots
  ..@ name  : chr "Constant Force Mortality"
  ..@ states: chr [1:2] "Alive" "Dead"
  ..@ Qxt   :List of 4
  .. ..$ :function (t = 0, x = 0, muxt = list(function (t = 0, x = 0, mu = 0.05) 
mu))  
  .. ..$ :function (t = 0, x = 0, mu = 0.05)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 144 26 145 21 26 21 144 145
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 146 26 147 24 26 24 146 147
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, muxt = list(function (t = 0, x = 0) 
0 * t))  
  .. ..- attr(*, "dim")= int [1:2] 2 2
}}
\usage{
cfm
}
\description{
An alive dead model with a constant force of mortality (i.e. exponential distribution).
}
\keyword{datasets}

