% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msm-package.R
\name{cmodel.object}
\alias{cmodel.object}
\title{Developer documentation: censoring model object}
\value{
\item{ncens}{The number of distinct values used for censored
observations in the \code{state} data supplied to \code{\link{msm}}.}
\item{censor}{A vector of length \code{ncens}, giving the labels used for
censored states in the data.} \item{states}{A vector obtained by
\code{unlist()}ing a list with \code{ncens} elements, each giving the set of
true states that an observation with this label could be.}
\item{index}{Index into \code{states} for the first state corresponding to
each \code{censor}, plus an extra \code{length(states)+1}.}
}
\description{
A list giving information about censored states, their labels in the data
and what true states they represent.
}
\seealso{
\code{\link{msm.object}}.
}
