% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{common}
\alias{common}
\title{Common Shock Model [msm]}
\format{\preformatted{Formal class 'msm' [package "MSM"] with 3 slots
  ..@ name  : chr "Common Shock Model"
  ..@ states: chr [1:4] "Both" "Husband" "Wife" "Dead"
  ..@ Qxt   :List of 16
  .. ..$ :function (t = 0, x = 0, y = 0, muxt = list(function (t = 0, x = 0, y = 0, 
    A = 1e-04, B = 3e-04, c = 1.075) 
A + B * c^(y + t), function (t = 0, x = 0, y = 0, A = 1e-04, D = 0.00035, 
    c = 1.075) 
A + D * c^(x + t), function (t = 0, x = 0, y = 0) 
5 * 10^(-5) * t^0))  
  .. ..$ :function (t = 0, x = 0, y = 0, A = 1e-04, B = 3e-04, c = 1.075)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 104 27 105 33 27 33 104 105
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, y = 0, A = 1e-04, D = 0.00035, c = 1.075)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 106 27 107 33 27 33 106 107
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, y = 0)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 108 27 109 37 27 37 108 109
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, y = 0)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 110 27 111 25 27 25 110 111
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, y = 0, muxt = list(function (t = 0, x = 0, y = 0) 
0 * t, function (t = 0, x = 0, y = 0) 
0 * t, function (t = 0, x = 0, y = 0) 
0 * t))  
  .. ..$ :function (t = 0, x = 0, y = 0)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 114 27 115 25 27 25 114 115
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, y = 0)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 116 27 117 25 27 25 116 117
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, y = 0)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 118 27 119 25 27 25 118 119
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, y = 0)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 120 27 121 25 27 25 120 121
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, y = 0, muxt = list(function (t = 0, x = 0, y = 0) 
0 * t, function (t = 0, x = 0, y = 0) 
0 * t, function (t = 0, x = 0, y = 0) 
0 * t))  
  .. ..$ :function (t = 0, x = 0, y = 0)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 124 27 125 25 27 25 124 125
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, y = 0)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 126 27 127 25 27 25 126 127
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, y = 0)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 128 27 129 25 27 25 128 129
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, y = 0)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 130 27 131 25 27 25 130 131
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, y = 0, muxt = list(function (t = 0, x = 0, y = 0) 
0 * t, function (t = 0, x = 0, y = 0) 
0 * t, function (t = 0, x = 0, y = 0) 
0 * t))  
  .. ..- attr(*, "dim")= int [1:2] 4 4
}}
\usage{
common
}
\description{
A multiple state model with states Both Alive, Husband Alive Only, Wife Alive Only, Both Dead
}
\details{
This model is shown in Exercise 8.5 of Actuarial Mathematics for Life Contingent Risks (1st edition)
}
\keyword{datasets}

