% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{crit}
\alias{crit}
\title{Critical Illness Model [msm]}
\format{\preformatted{Formal class 'msm' [package "MSM"] with 3 slots
  ..@ name  : chr "Critical Illness Model"
  ..@ states: chr [1:4] "Healthy" "Sick" "Dead" "Critical"
  ..@ Qxt   :List of 16
  .. ..$ :function (t = 0, x = 0, muxt = list(function (t = 0, x = 0, a1 = 4e-04, 
    b1 = 3.5e-06, c1 = 0.14) 
a1 + b1 * exp(c1 * (x + t)), function (t = 0, x = 0, a2 = 5e-04, b2 = 7.6e-05, 
    c2 = 0.09) 
a2 + b2 * exp(c2 * (x + t)), function (t = 0, x = 0, a1 = 4e-04, b1 = 3.5e-06, 
    c1 = 0.14) 
0.05 * (a1 + b1 * exp(c1 * (x + t)))))  
  .. ..$ :function (t = 0, x = 0, a1 = 4e-04, b1 = 3.5e-06, c1 = 0.14)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 63 26 64 43 26 43 63 64
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, a2 = 5e-04, b2 = 7.6e-05, c2 = 0.09)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 65 26 66 43 26 43 65 66
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, a1 = 4e-04, b1 = 3.5e-06, c1 = 0.14)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 67 26 68 52 26 52 67 68
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, a1 = 4e-04, b1 = 3.5e-06, c1 = 0.14)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 69 26 70 51 26 51 69 70
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, muxt = list(function (t = 0, x = 0, a1 = 4e-04, 
    b1 = 3.5e-06, c1 = 0.14) 
0.1 * (a1 + b1 * exp(c1 * (x + t))), function (t = 0, x = 0, a2 = 5e-04, 
    b2 = 7.6e-05, c2 = 0.09) 
a2 + b2 * exp(c2 * (x + t)), function (t = 0, x = 0, a1 = 4e-04, b1 = 3.5e-06, 
    c1 = 0.14) 
0.05 * (a1 + b1 * exp(c1 * (x + t)))))  
  .. ..$ :function (t = 0, x = 0, a2 = 5e-04, b2 = 7.6e-05, c2 = 0.09)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 73 26 74 43 26 43 73 74
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, a1 = 4e-04, b1 = 3.5e-06, c1 = 0.14)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 75 26 76 51 26 51 75 76
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 77 26 78 24 26 24 77 78
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 79 26 80 24 26 24 79 80
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, muxt = list(function (t = 0, x = 0) 
0 * t, function (t = 0, x = 0) 
0 * t, function (t = 0, x = 0) 
0 * t))  
  .. ..$ :function (t = 0, x = 0)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 83 26 84 24 26 24 83 84
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 85 26 86 24 26 24 85 86
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 87 26 88 24 26 24 87 88
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, a2 = 5e-04, b2 = 7.6e-05, c2 = 0.09)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 89 26 90 50 26 50 89 90
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, muxt = list(function (t = 0, x = 0) 
0 * t, function (t = 0, x = 0) 
0 * t, function (t = 0, x = 0, a2 = 5e-04, b2 = 7.6e-05, c2 = 0.09) 
1.2 * (a2 + b2 * exp(c2 * (x + t)))))  
  .. ..- attr(*, "dim")= int [1:2] 4 4
}}
\usage{
crit
}
\description{
A multiple state model with states Healthy, Sick, Dead, Critical
}
\details{
This model is shown in Exercise 8.2 of Actuarial Mathematics for Life Contingent Risks (1st edition)
}
\keyword{datasets}

