% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{dm}
\alias{dm}
\title{Disability Model [msm]}
\format{\preformatted{Formal class 'msm' [package "MSM"] with 3 slots
  ..@ name  : chr "Disability model"
  ..@ states: chr [1:3] "Alive" "Disabled" "Dead"
  ..@ Qxt   :List of 9
  .. ..$ :function (t = 0, x = 0, muxt = list(function (t = 0, x = 0, a1 = 4e-04, 
    b1 = 3.4674e-06, c1 = 0.138155) 
a1 + b1 * exp(c1 * (x + t)), function (t = 0, x = 0, a2 = 5e-04, b2 = 7.5858e-05, 
    c2 = 0.087498) 
a2 + b2 * exp(c2 * (x + t))))  
  .. ..$ :function (t = 0, x = 0, a1 = 4e-04, b1 = 3.4674e-06, c1 = 0.138155)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 9 25 10 41 25 41 9 10
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, a2 = 5e-04, b2 = 7.5858e-05, c2 = 0.087498)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 11 25 12 39 25 39 11 12
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, a1 = 4e-04, b1 = 3.4674e-06, c1 = 0.138155)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 13 25 14 51 25 51 13 14
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, muxt = list(function (t = 0, x = 0, a1 = 4e-04, 
    b1 = 3.4674e-06, c1 = 0.138155) 
0.1 * (a1 + b1 * exp(c1 * (x + t))), function (t = 0, x = 0, a2 = 5e-04, 
    b2 = 7.5858e-05, c2 = 0.087498) 
a2 + b2 * exp(c2 * (x + t))))  
  .. ..$ :function (t = 0, x = 0, a2 = 5e-04, b2 = 7.5858e-05, c2 = 0.087498)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 17 25 18 39 25 39 17 18
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 19 25 20 23 25 23 19 20
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0)  
  .. .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 21 25 22 23 25 23 21 22
  .. .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000139d72e0> 
  .. ..$ :function (t = 0, x = 0, muxt = list(function (t = 0, x = 0) 
0 * t, function (t = 0, x = 0) 
0 * t))  
  .. ..- attr(*, "dim")= int [1:2] 3 3
}}
\usage{
dm
}
\description{
A multiple state model with states Alive, Disabled and Dead
}
\keyword{datasets}

