% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mdm.R
\docType{class}
\name{msm-class}
\alias{msm-class}
\title{Multiple State Model Class}
\arguments{
\item{representation}{Each multiple state model contains a name, state names and a transition matrix}
}
\description{
Multiple State Model Class
}
\examples{
udd <- new("msm", name="Uniform Distribution Deaths",
    states = c("Alive", "Dead"),
    Qxt = Qxt.diag( matrix(c(
       uxt00 = function(t=0, x=0)
           0
       , uxt01 = function(t=0, x=0, w=120)
           1/(w-x-t)
       , uxt10 = function(t=0, x=0)
           0 * t
       , uxt11 = function(t=0, x=0)
           0
    ), 2, 2)))
print(udd)
}

