% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Msm_A.R
\name{Msm_A}
\alias{Msm_A}
\title{Transition Probability Matrix for \code{\link{Msm}}(k) Model.}
\usage{
Msm_A(b, gamma.kbar, kbar)
}
\arguments{
\item{b}{is the  growth rate of the switching probabilities of the volatility components and
must be in the range (1, inf).}

\item{gamma.kbar}{is the transition probability of the highest frequency component
and must be in the range (0,1).}

\item{kbar}{is the number of frequency components in the \code{\link{Msm}}(k) model.}
}
\value{
a \eqn{2^k} by \eqn{2^k} transition matrix
}
\description{
Calculates transition matrix for \code{\link{Msm}}(k) model. The calculation
 exploits the independence of the volatility components. Each \eqn{M_k} has its own
 \eqn{2 x 2} transition matrix \eqn{A_k}. It can be shown that the transition matrix
 for \eqn{M=[M_1;M_2,..M_{kbar}]} is equal to the kronecker product of all the
 transition matrices for all the \eqn{M_k}.
}
\examples{
Msm_A(10.91, 0.11, 2)

}

