% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Msm_clustermat.R
\name{Msm_clustermat}
\alias{Msm_clustermat}
\title{Volatility State Matrix for \code{\link{Msm}}(k) model.}
\usage{
Msm_clustermat(m0, kbar)
}
\arguments{
\item{m0}{is the value of each volatility components and must be in the range (1, 2].}

\item{kbar}{is the number of frequency components in the \code{\link{Msm}}(k) model.}
}
\value{
a \eqn{2^k} by \eqn{k} grid of volatility components.
}
\description{
Calculates the matrix of volatility components for \code{\link{Msm}}(k) model.
}
\examples{
Msm_clustermat(1.5, 2)

}

