% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Msm_decompose.R
\name{Msm_decompose}
\alias{Msm_decompose}
\title{Decomposes returns into volatility components.}
\usage{
Msm_decompose(msmobject)
}
\arguments{
\item{msmobject}{is an object of class "msmmodel" returned by \code{\link{Msm}}.}
}
\value{
An object of class msmcomp containing the \eqn{T} by \eqn{k} matrix of volatility components, \eqn{E(M)}.
 The returned object can be coarsed into a matrix with \code{\link{as.matrix}} function.
 Use \code{\link{plot.msmmodel}} to plot the volatility components.
}
\description{
Decomposes returns into \eqn{k>1} volatility components.
}
\examples{
data("calvet2004data")
ret <- na.omit(as.matrix(dat$caret))*100
fit2 <- Msm(ret, kbar=2, n.vol=252, nw.lag=2)
volcomp <- Msm_decompose(fit2)
plot(volcomp)


}
\author{
Waleem Alausa, \email{alausa.babs@gmail.com}
}
\keyword{Markov}
\keyword{Multifractal,}
\keyword{Switching}
\keyword{clustering}
\keyword{volatility}
\keyword{volatility,}

