% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Msm_grad.R
\name{Msm_grad}
\alias{Msm_grad}
\title{Calculates numerical gradient for MSM(k) model.}
\usage{
Msm_grad(para, kbar, ret, n.vol)
}
\arguments{
\item{para}{is a vector of parameters returned by \code{\link{Msm}}.}

\item{kbar}{is the number of frequency components in the \code{\link{Msm}}(k) model.}

\item{ret}{is a column matrix of returns.}

\item{n.vol}{is the number of trading days in a year, if data is on a daily frequency. Default is 252.}
}
\value{
a \eqn{N=nrow(ret)} by \eqn{4} matrix of numerical derivatives.
}
\description{
Calculates numerical gradient for \code{\link{Msm}}(k) model.
}

