% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Msm_hessian_2_sided.R
\name{Msm_hesssian_2_sided}
\alias{Msm_hesssian_2_sided}
\title{Finite Difference Hessian for \code{\link{Msm}}(k) model.}
\usage{
Msm_hesssian_2_sided(para, kbar, ret, n.vol)
}
\arguments{
\item{para}{is a vector of parameters returned by \code{\link{Msm}}.}

\item{kbar}{is the number of frequency components in the \code{\link{Msm}}(k) model.}

\item{ret}{is a column matrix of returns.}

\item{n.vol}{is the number of trading days in a year, if data is on a daily frequency. Default is 252.}
}
\value{
a \eqn{4 x 4} matrix of numerical hessians.
}
\description{
Calculates 2-sided finite difference hessian for \code{\link{Msm}}(k) model.
}

