% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Msm_likelihood2.R
\name{Msm_likelihood2}
\alias{Msm_likelihood2}
\title{Log-Likelihood for \code{\link{Msm}}(k) model.}
\usage{
Msm_likelihood2(para, kbar, dat, n.vol)
}
\arguments{
\item{para}{is a vector of parameters [m0, b, gamma_k, sigma].}

\item{kbar}{is the number of frequency components in the \code{\link{Msm}}(k) model.}

\item{dat}{is a column matrix of returns.}

\item{n.vol}{is the number of trading days in a year, if data is on a daily frequency. Default is 252.
If data is monthly, then set n.vol to 12. If data is quarterly, then n.vol should be 4.}
}
\value{
a list consisting of:
\item{LL}{sum of loglikelihood values at optimum}
\item{LLs}{a vector of loglikelihood values at optimum}
\item{filtered}{a matrix of filtered probabilities}
\item{A}{the estimated transition matrix}
\item{g.m}{the estimated state values}
}
\description{
Calculates log-likelihood and filtered probability values for \code{\link{Msm}}(k) model.
}
\examples{
data("calvet2004data")
ret <- na.omit(as.matrix(calvet2004data$caret))*100
vol <- Msm_likelihood2(c(1.556, 10.92, 0.109, 0.278*sqrt(252)), 2, ret, n.vol=252)

}

