% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Msm_ll2.R
\name{Msm_ll2}
\alias{Msm_ll2}
\title{Calculates the sum of the log-likelihood values for MSM(k) model.}
\usage{
Msm_ll2(para, kbar, dat, n.vol)
}
\arguments{
\item{para}{is a vector of parameters returned by \code{\link{Msm}}.}

\item{kbar}{is the number of frequency components in the \code{\link{Msm}}(k) model.}

\item{dat}{is a column matrix of returns.}

\item{n.vol}{is the number of trading days in a year, if data is on a daily frequency. Default is 252.}
}
\value{
sum of the log-likelihood values
}
\description{
Calculates the sum of the log-likelihood values for \code{\link{Msm}}(k) model.
}
\examples{
data("calvet2004data")
ret <- na.omit(as.matrix(dat$caret))*100
Msm_ll2(c(1.556, 10.92, 0.109, 0.278*sqrt(252)), 2, ret, n.vol=252)

}

