% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Msm_parameter_check.R
\name{Msm_parameter_check}
\alias{Msm_parameter_check}
\title{Performs boundary checks on Msm model parameters.}
\usage{
Msm_parameter_check(dat, kbar, x0)
}
\arguments{
\item{dat}{is a column matrix/ dataframe of returns.}

\item{kbar}{is the number of frequency components in the \code{\link{Msm}}(k) model.}

\item{x0}{is the initial parameter values passed to Msm.}
}
\value{
a list consisting of:
\item{dat}{is a column matrix of returns}
\item{kbar}{number of frequency components}
\item{start.value}{is the initial parameter values passed to Msm}
\item{lb}{lower bound on parameters}
\item{ub}{lower bound on parameters}
}
\description{
Performs boundary checks on Msm model parameter values.
}
\examples{
data("calvet2004data")
ret <- na.omit(as.matrix(dat$caret))*100
Msm_parameter_check (ret,2,c(1.5,10,.1,4))

}

