% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Msm_predict.R
\name{Msm_predict}
\alias{Msm_predict}
\title{Prediction for \code{\link{Msm}} Model.}
\usage{
Msm_predict(g.m, sigma, n, P, A, h = NULL)
}
\arguments{
\item{g.m}{is is a vector of possible msm states.}

\item{sigma}{is the unconditional volatility returned by \code{\link{Msm}}(k) model.}

\item{n}{is the number of trading days in a year, if data is on a daily frequency. Default is 252.
If data is monthly, then set n.vol to 12. If data is quarterly, then n.vol should be 4.}

\item{P}{is a matrix of smoothed/filtered state probabilities.}

\item{A}{is the transition matrix returned by Msm.}

\item{h}{is an optional parameter. When h is null, the fitted volatility values are returned.
When h>0, then an h-step ahead volatility forecast is returned.}
}
\value{
a list consisting of:
\item{vol}{fitted/forecasted volatility values}
\item{LLs}{fitted/forecasted variance values}
}
\description{
Performs prediction after fitting an \code{\link{Msm}}(k) model.
}
\examples{
data("calvet2004data")
ret <- na.omit(as.matrix(calvet2004data$caret))*100
fit <- Msm(ret, kbar=2, n.vol=252, nw.lag=2)
vol <- Msm_predict(fit$g.m, fit$para[4]*sqrt(252), 252, fit$filtered, fit$A, 5)
}

