% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{MSM_standalone}
\alias{MSM_standalone}
\title{Standalone MSM}
\usage{
MSM_standalone(
  title = NULL,
  num_items = 10L,
  with_id = FALSE,
  with_welcome = TRUE,
  with_training = TRUE,
  type = "PART1",
  admin_password = "conifer",
  researcher_email = "klaus.frieler@ae.mpg.de",
  languages = c("en", "de"),
  dict = MSM::MSM_dict,
  audio_dir = "https://s3-eu-west-1.amazonaws.com/media.dots.org/stimuli/MSM",
  validate_id = "auto",
  ...
)
}
\arguments{
\item{title}{(Scalar character) Title to display during testing.}

\item{num_items}{(Scalar integer) Number of items to be adminstered.}

\item{with_id}{(Logical scalar) Whether to show a ID page.}

\item{with_welcome}{(Logical scalar) Whether to show a welcome page.}

\item{with_training}{(Logical scalar) Whether to include the training phase.}

\item{type}{(character scalar) Which version to show (either PART1 or PART2)}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}),
and German (\code{"DE"}).
The first language is selected by default}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{audio_dir}{(URL) The top level URL for audio stimuli}

\item{validate_id}{(Character scalar or closure) Function for validating IDs or string "auto" for default validation
which means ID should consist only of  alphanumeric characters.}

\item{...}{Further arguments to be passed to \code{\link{MSM}()}.}
}
\description{
This function launches a standalone testing session for the MSM
This can be used for data collection, either in the laboratory or online.
}
