\name{print.msm}
\title{Print a fitted msm model object}
\alias{print.msm}
\description{
  Print a fitted msm model object
}
\usage{
\method{print}{msm}(x,\dots)
}
\arguments{
  \item{x}{Output from \code{\link{msm}}, representing a fitted
    multi-state model object.}
  \item{...}{Other arguments to be passed to \code{\link{format}}.}
}
\seealso{
  \code{\link{printnew.msm}}
}
\details{
See \code{\link{printnew.msm}} for a better and cleaner output format,
and an explanation of the change.
}
\author{C. H. Jackson \email{chris.jackson@mrc-bsu.cam.ac.uk}}
\keyword{models}
