% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{printold.msm}
\alias{printold.msm}
\title{Print a fitted msm model object}
\usage{
printold.msm(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{msm}}, representing a fitted multi-state
model object.}

\item{...}{Other arguments to be passed to \code{\link{format}}.}
}
\description{
Print a fitted msm model object (in old format, from msm 1.3.1 and earlier)
}
\details{
See \code{\link{print.msm}} for a better and cleaner output format, and an
explanation of the change.
}
\seealso{
\code{\link{print.msm}}
}
\author{
C. H. Jackson \email{chris.jackson@mrc-bsu.cam.ac.uk}
}
\keyword{models}
