% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mdm.R
\name{Qxt.diag}
\alias{Qxt.diag}
\title{Diagonal of transition matrix}
\usage{
Qxt.diag(Qxt.fns, zero_diag = FALSE, joint = FALSE)
}
\arguments{
\item{Qxt.fns}{A matrix of transition functions}

\item{zero_diag}{Indicate whether the diagonal should be all functions which return 0}

\item{joint}{Indicate whether the tranisition matrix is for joint lives (i.e. functions of x and y)}
}
\value{
A matrix of transition functions with updated diagonal entries
}
\description{
Generates functions for the diagonal of the transition matrix, Q
}

