% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.msm.est}
\alias{tidy.msm.est}
\title{Tidy the output of pmatrix.msm and similar functions}
\usage{
\method{tidy}{msm.est}(x, ...)
}
\arguments{
\item{x}{Output of \code{\link{qmatrix.msm}}, \code{\link{pmatrix.msm}}, 
\code{\link{ematrix.msm}}, \code{\link{pnext.msm}} or \code{\link{ppass.msm}},
 which all return objects of class \code{"msm.est"}.}

\item{...}{Further arguments (unused).}
}
\description{
This is the method for the generic `tidy` function that is 
used for tidying the output of \code{\link{qmatrix.msm}}, \code{\link{pmatrix.msm}}, 
\code{\link{ematrix.msm}}, \code{\link{pnext.msm}} or \code{\link{ppass.msm}}.
 This should be called as 
\code{tidy()}, not \code{tidy.msm.est()} or \code{tidy.qmatrix()} or anything else.
}
