% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.msm.prevalence}
\alias{tidy.msm.prevalence}
\title{Tidy the output of prevalence.msm}
\usage{
\method{tidy}{msm.prevalence}(x, ...)
}
\arguments{
\item{x}{Output of \code{\link{prevalence.msm}}.}

\item{...}{Further arguments (unused).}
}
\value{
A tibble with one row per combination of output type (count or percentage)
and state, and columns for observed value, expected value and confidence 
limits for the expected value (if available).
}
\description{
Note this should be called as \code{tidy()} not \code{tidy.msm.prevalence()} or anything else, as this is 
a method for the generic \code{tidy()} function.
}
