% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mdm.R
\name{tPx}
\alias{tPx}
\title{Transition Probabilities}
\usage{
tPx(object, t = 0, x = 0, h = 1/12, all = FALSE, joint = FALSE,
  y = 0)
}
\arguments{
\item{object}{An object of class msm}

\item{t}{The number of years forward}

\item{x}{The age of the individual.}

\item{h}{The time step. By default 1/12 is used for the approximation}

\item{all}{Indicates whether the probabilities at each time point should be returned or whether only the final probability matrix should be returned (at time t)}

\item{joint}{Indicate whether there are two lives, x and y that should should be used in the transition matrix functions}

\item{y}{The age of the second individual if the model is joint}
}
\value{
A matrix of transition probabilities
}
\description{
Probability of transitioning from one state to another within t years. Approximation based on the Kolmogorov forward equations.
}

