\name{aneur}
\alias{aneur}
\title{Aortic aneurysm progression data}
\description{
  This dataset contains a list of transitions between
  grades of aortic aneurysms,
}
\usage{data(aneur)}

\format{
  A data frame containing 2662 rows. Columns correspond to
  'starting state', 'end state', 'time interval', 'age at beginning of
  interval'. }

\source{
  The Chichester, U.K. randomised controlled trial of screening for
  abdominal aortic aneurysms by ultrasonography.}

\references{
  Couto, E. and Duffy, S. W. and Ashton, H. A. and Walker, N. M.
  and Myles, J. P. and Scott, R. A. P. and Thompson, S. G. (2002)
  \emph{Probabilities of progression of aortic aneurysms: estimates
    and implications for screening policy}
  Journal of Medical Screening, in press.
}

\keyword{datasets}
