% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msm-package.R
\docType{package}
\name{msm-package}
\alias{msm-package}
\alias{_PACKAGE}
\title{Multi-State Markov and Hidden Markov Models in Continuous Time}
\description{
msm: Functions for fitting continuous-time Markov and hidden Markov
multi-state models to longitudinal data.  Designed for processes
observed at arbitrary times in continuous time (intermittently
observed or panel data) but some other observation schemes are
supported. Both Markov transition rates and the hidden Markov
output process can be modelled in terms of covariates, which may be
constant or piecewise-constant in time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chjackson/msm}
  \item \url{https://chjackson.github.io/msm/}
  \item Report bugs at \url{https://github.com/chjackson/msm/issues}
}

}
