% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSM.R
\name{MSM}
\alias{MSM}
\title{MSM}
\usage{
MSM(
  with_welcome = TRUE,
  with_training = TRUE,
  with_finish = TRUE,
  label = "MSM",
  audio_dir = "https://s3-eu-west-1.amazonaws.com/media.dots.org/stimuli/MSM",
  dict = MSM::MSM_dict,
  ...
)
}
\arguments{
\item{with_welcome}{(Logical scalar) Whether to show a welcome page.}

\item{with_training}{(Logical scalar) Whether to include the training phase.}

\item{with_finish}{(Logical scalar) Whether to show a finished page.}

\item{label}{(Character scalar) Label to give the MSM results in the output file.}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{num_items}{(Integer scalar) Number of items in the test.}
}
\description{
This function defines a MSM  module for incorporation into a
psychTestR timeline.
Use this function if you want to include the MSM in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For demoing the MSM, consider using \code{\link{MSM_demo}()}.
For a standalone implementation of the MSM,
consider using \code{\link{MSM_standalone}()}.
}
