% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Msm.R
\name{Msm}
\alias{Msm}
\title{Fits an \code{\link{Msm}}(k) model to returns.}
\usage{
Msm(ret, kbar = 1, n.vol = 252, para0 = NULL, nw.lag = 0)
}
\arguments{
\item{ret}{is a column matrix of returns.}

\item{kbar}{is the number of frequency components in the \code{\link{Msm}}(k) model. Default is 1.}

\item{n.vol}{is the number of trading days in a year, if data is on a daily frequency. Default is 252.
If data is monthly, then set n.vol to 12. If data is quarterly, then n.vol should be 4.}

\item{para0}{is the initial parameter values for optimization. Default is NULL.}

\item{nw.lag}{is the number of lags to use for calculating robust Newey-West standard errors. Default is 2.}
}
\value{
a list consisting of:
\item{LL}{sum of loglikelihood values at optimum}
\item{LLs}{a vector of loglikelihood values at optimum}
\item{filtered}{a matrix of filtered probabilities}
\item{A}{the estimated transition matrix}
\item{g.m}{the estimated state values}
\item{optim.msg}{optimization message}
\item{optim.convergence}{optimization convergence indicator}
\item{optim.iter}{number of iterations}
\item{para]}{estimated parameter vector}
\item{se}{robust standard errors}
}
\description{
Estimates the parameters of an \code{\link{Msm}}(k) model for a given vector of returns.
}
\note{
Use the \code{\link{summary.msmmodel}} function to print summary results of the MSM model.
Use \code{\link{plot.msmmodel}} function to plot the fitted conditional volatilities, or plot(msmmodel, what = "volsq") to plot the
 fitted  conditional variance. Use \code{\link{predict.msmmodel}} to get the predicted conditional volatility and variance.
}
\examples{
data("calvet2004data")
ret <- na.omit(as.matrix(dat$caret))*100
fit <- Msm(ret)
fit2 <- Msm(ret, kbar=2, n.vol=252, nw.lag=2)
summary(fit2)
plot(fit2)

}
\author{
Waleem Alausa, \email{alausa.babs@gmail.com}
}
\keyword{Markov}
\keyword{Multifractal,}
\keyword{Switching}
\keyword{clustering}
\keyword{volatility}
\keyword{volatility,}

