% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Msm_likelihood2.R
\name{Msm_likelihood2}
\alias{Msm_likelihood2}
\title{Calculates the log-likelihood values for MSM(k) model.}
\usage{
Msm_likelihood2(para, kbar, dat, n.vol)

Msm_likelihood2(para, kbar, dat, n.vol)
}
\arguments{
\item{para}{is a vector of parameters returned by \code{\link{Msm}}.}

\item{kbar}{is the number of frequency components in the \code{\link{Msm}}(k) model.}

\item{dat}{is a column matrix of returns.}

\item{n.vol}{is the number of trading days in a year, if data is on a daily frequency. Default is 252.}

\item{para}{is a vector of parameters returned by \code{\link{Msm}}.}

\item{kbar}{is the number of frequency components in the \code{\link{Msm}}(k) model.}

\item{dat}{is a column matrix of returns.}

\item{n.vol}{is the number of trading days in a year, if data is on a daily frequency. Default is 252.}
}
\value{
a list consisting of:
\item{LL}{sum of loglikelihood values at optimum}
\item{LLs}{a vector of loglikelihood values at optimum}
\item{filtered}{a matrix of filtered probabilities}
\item{A}{the estimated transition matrix}
\item{g.m}{the estimated state values}

a list consisting of:
\item{LL}{sum of loglikelihood values at optimum}
\item{LLs}{a vector of loglikelihood values at optimum}
\item{filtered}{a matrix of filtered probabilities}
\item{A}{the estimated transition matrix}
\item{g.m}{the estimated state values}
}
\description{
Calculates the log-likelihood values for \code{\link{Msm}}(k) model.

Calculates the log-likelihood values for \code{\link{Msm}}(k) model.
}
\examples{
data("calvet2004data")
ret <- na.omit(as.matrix(dat$caret))*100
Msm_likelihood2(c(1.556, 10.92, 0.109, 0.278*sqrt(252)), 2, ret, n.vol=252)

data("calvet2004data")
ret <- na.omit(as.matrix(dat$caret))*100
Msm_likelihood2(c(1.556, 10.92, 0.109, 0.278*sqrt(252)), 2, ret, n.vol=252)

}

