\name{printold.msm}
\title{Print a fitted msm model object}
\alias{printold.msm}
\description{
  Print a fitted msm model object (in old format, from msm 1.3.1 and earlier)
}
\usage{
printold.msm(x,\dots)
}
\arguments{
  \item{x}{Output from \code{\link{msm}}, representing a fitted
    multi-state model object.}
  \item{...}{Other arguments to be passed to \code{\link{format}}.}
}
\seealso{
  \code{\link{print.msm}}
}
\details{
See \code{\link{print.msm}} for a better and cleaner output format,
and an explanation of the change.
}
\author{C. H. Jackson \email{chris.jackson@mrc-bsu.cam.ac.uk}}
\keyword{models}
