% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.msm.est}
\alias{tidy.msm.est}
\title{Tidy the output of a msm.est object, a list of matrices with estimates, L and U components}
\usage{
\method{tidy}{msm.est}(x, ...)
}
\arguments{
\item{x}{Output of \code{\link{qmatrix.msm}}, \code{\link{pmatrix.msm}}, 
\code{\link{ematrix.msm}}, \code{\link{pnext.msm}} or \code{\link{ppass.msm}},
 which all return objects of class `"msm.est"`.}

\item{...}{Further arguments (unused).}
}
\description{
This is the method for the generic `tidy` function that is 
used for tidying the output of `qmatrix.msm`, `pmatrix.msm`, `ematrix.msm`,
`pnext.msm` and `ppass.msm`.  This should be called as 
`tidy()`, not `tidy.msm.est()` or `tidy.qmatrix()` or anything else.
}
