% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.msm.estbystate}
\alias{tidy.msm.estbystate}
\title{Tidy the output of totlos.msm and similar functions}
\usage{
\method{tidy}{msm.estbystate}(x, ...)
}
\arguments{
\item{x}{Output of \code{\link{totlos.msm}}, \code{\link{envisits.msm}}
or \code{\link{efpt.msm}}, which return objects of class \code{"msm.estbystate"}.}

\item{...}{Further arguments (unused).}
}
\value{
A tibble with one row per state, and columns for the estimate, and
 confidence intervals if available.
}
\description{
Note this should be called as \code{tidy()} not \code{tidy.msm.totlos()} or anything else, as this is 
a method for the generic \code{tidy()} function.
}
